@@ -20,7 +20,7 @@ if("${CMAKE_VERSION}" VERSION_GREATER 2.8.11)
 	cmake_policy(SET CMP0020 OLD)
 endif()
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
